--- a/modules/platforms/dotnet/Apache.Ignite.Core/Impl/Binary/BinaryUtils.cs
+++ b/modules/platforms/dotnet/Apache.Ignite.Core/Impl/Binary/BinaryUtils.cs
@@ -109,9 +109,9 @@ namespace Apache.Ignite.Core.Impl.Binary
         /** String mode. */
         public static readonly bool UseStringSerializationVer2 =
             (Environment.GetEnvironmentVariable(IgniteBinaryMarshallerUseStringSerializationVer2) ?? "false") == "true";
-        
+
         /** Cached maps of enum members per type. */
-        private static readonly CopyOnWriteConcurrentDictionary<Type, Dictionary<string, int>> EnumValues = 
+        private static readonly CopyOnWriteConcurrentDictionary<Type, Dictionary<string, int>> EnumValues =
             new CopyOnWriteConcurrentDictionary<Type, Dictionary<string, int>>();
 
         /// <summary>
@@ -925,7 +925,7 @@ namespace Apache.Ignite.Core.Impl.Binary
         }
 
         /// <summary>
-        /// Writes a guid with bitwise conversion, assuming that <see cref="Guid"/> 
+        /// Writes a guid with bitwise conversion, assuming that <see cref="Guid"/>
         /// is laid out in memory sequentially and without gaps between fields.
         /// </summary>
         /// <param name="val">The value.</param>
@@ -973,7 +973,7 @@ namespace Apache.Ignite.Core.Impl.Binary
         }
 
         /// <summary>
-        /// Reads a guid with bitwise conversion, assuming that <see cref="Guid"/> 
+        /// Reads a guid with bitwise conversion, assuming that <see cref="Guid"/>
         /// is laid out in memory sequentially and without gaps between fields.
         /// </summary>
         /// <param name="stream">The stream.</param>
@@ -1102,8 +1102,8 @@ namespace Apache.Ignite.Core.Impl.Binary
         /// </summary>
         public static int GetArrayElementTypeId(Type elemType, Marshaller marsh)
         {
-            return elemType == typeof(object) 
-                ? ObjTypeId 
+            return elemType == typeof(object)
+                ? ObjTypeId
                 : marsh.GetDescriptor(elemType).TypeId;
         }
 
@@ -1167,6 +1167,38 @@ namespace Apache.Ignite.Core.Impl.Binary
                 vals[i] = ctx.Deserialize<T>();
 
             return vals;
+        }
+
+        /// <summary>
+        /// Read string array.
+        /// </summary>
+        /// <param name="stream">Stream</param>
+        /// <returns>String array.</returns>
+        public static string[] ReadStringArray(IBinaryStream stream)
+        {
+            var len = stream.ReadInt();
+            var res = new string[len];
+
+            for (var i = 0; i < len; i++)
+                res[i] = stream.ReadByte() == HdrNull ? null : ReadString(stream);
+
+            return res;
+        }
+
+        /// <summary>
+        /// Read string array.
+        /// </summary>
+        /// <param name="stream">Stream</param>
+        /// <returns>String array.</returns>
+        public static Guid?[] ReadGuidArray(IBinaryStream stream)
+        {
+            var len = stream.ReadInt();
+            var res = new Guid?[len];
+
+            for (var i = 0; i < len; i++)
+                res[i] = stream.ReadByte() == HdrNull ? (Guid?) null : ReadGuid(stream);
+
+            return res;
         }
 
         /// <summary>
@@ -1595,7 +1627,7 @@ namespace Apache.Ignite.Core.Impl.Binary
             err = null;
 
             if (reader.ReadBoolean()) // success indication
-                return reader.ReadObject<object>(); 
+                return reader.ReadObject<object>();
 
             err = reader.ReadBoolean() // native error indication
                 ? reader.ReadObject<object>()
@@ -1745,7 +1777,7 @@ namespace Apache.Ignite.Core.Impl.Binary
 
             var enumType = Enum.GetUnderlyingType(type);
 
-            return enumType == typeof(int) || enumType == typeof(byte) || enumType == typeof(sbyte) 
+            return enumType == typeof(int) || enumType == typeof(byte) || enumType == typeof(sbyte)
                 || enumType == typeof(short) || enumType == typeof(ushort) || enumType == typeof(uint);
         }
 
@@ -1762,7 +1794,7 @@ namespace Apache.Ignite.Core.Impl.Binary
 
             return TimeSpan.FromMilliseconds(ms);
         }
-        
+
         /// <summary>
         /// Gets the enum values.
         /// </summary>
@@ -1770,7 +1802,7 @@ namespace Apache.Ignite.Core.Impl.Binary
         {
             Debug.Assert(enumType != null);
             Debug.Assert(enumType.IsEnum);
-            
+
             Dictionary<string,int> res;
             if (EnumValues.TryGetValue(enumType, out res))
             {
@@ -1791,10 +1823,10 @@ namespace Apache.Ignite.Core.Impl.Binary
             }
 
             EnumValues.Set(enumType, res);
-            
+
             return res;
         }
-        
+
         /// <summary>
         /// Gets the enum value as int.
         /// </summary>