… modules/platforms/dotnet/Apache.Ignite.BenchmarkDotNet/Apache.Ignite.BenchmarkDotNet.snk
… modules/platforms/dotnet/Apache.Ignite.Core/Impl/Binary/BinarySystemHandlers.cs
… modules/platforms/dotnet/Apache.Ignite.Core/Properties/AssemblyInfo.cs