Replace dictionary+interfaces with a simple switch statement in `BinarySystemHandlers.TryReadSystemType`:
* Less code
* Simpler
* Faster

Benchmark added: `BinarySystemTypeReadBenchmark`; results on Core i7-9700K, Ubuntu 20.04, .NET Core 3.1:

```
|        |     Mean |     Error |    StdDev |  Gen 0 | Gen 1 | Gen 2 | Allocated |
|------- |---------:|----------:|----------:|-------:|------:|------:|----------:|
| Before | 1.080 us | 0.0030 us | 0.0025 us | 0.0725 |     - |     - |     456 B |
|  After | 849.4 ns |   4.44 ns | 4.16 ns   | 0.0725 |     - |     - |     456 B |
```

(~20% improvement)